requirements.txt
torch>=2.0.0
numpy>=1.21.0
pandas>=1.3.0
scikit-learn>=0.24.2
matplotlib>=3.4.3
pillow>=8.3.1
torchvision>=0.15.0 